+++ solution/3000-3099/3013.Divide an Array Into Subarrays With Minimum Cost II/Solution.py
@@ -3,35 +3,41 @@ from sortedcontainers import SortedList
 
 class Solution:
     def minimumCost(self, nums: List[int], k: int, dist: int) -> int:
-        n = len(nums)
+        def l2r():
+            nonlocal s
+            x = l.pop()
+            s -= x
+            r.add(x)
 
-        sl = SortedList()
-        y = nums[0]
-        ans = float("inf")
-        i = 1
-        running_sum = 0
+        def r2l():
+            nonlocal s
+            x = r.pop(0)
+            l.add(x)
+            s += x
 
-        for j in range(1, n):
-            pos = bisect.bisect_left(sl, nums[j])
-            sl.add(nums[j])
-
-            if pos < k - 1:
-                running_sum += nums[j]
-                if len(sl) > k - 1:
-                    running_sum -= sl[k - 1]
-
-            while j - i > dist:
-                removed_pos = sl.index(nums[i])
-                removed_element = nums[i]
-                sl.remove(removed_element)
-
-                if removed_pos < k - 1:
-                    running_sum -= removed_element
-                    if len(sl) >= k - 1:
-                        running_sum += sl[k - 2]
-                i += 1
-
-            if j - i + 1 >= k - 1:
-                ans = min(ans, running_sum)
-
-        return ans + y
+        k -= 1
+        s = sum(nums[: dist + 2])
+        l = SortedList(nums[1 : dist + 2])
+        r = SortedList()
+        while len(l) > k:
+            l2r()
+        ans = s
+        for i in range(dist + 2, len(nums)):
+            x = nums[i - dist - 1]
+            if x in l:
+                l.remove(x)
+                s -= x
+            else:
+                r.remove(x)
+            y = nums[i]
+            if y < l[-1]:
+                l.add(y)
+                s += y
+            else:
+                r.add(y)
+            while len(l) < k:
+                r2l()
+            while len(l) > k:
+                l2r()
+            ans = min(ans, s)
+        return ans
